Add more typed helpers
We know the intent needed for some types of requests like picking a photo or taking a video, or choosing a contact, so we can expose these as async methods that are easily called without needing to construct the intent.

These helpers also pass back a more strongly typed version of the `ActivityResult`, with properties appropriate to retrieving result information.  For example, `ContactPickerActivityResult` has a `SelectedContactUri` property, and the `MediaPickerActivityResult` has a `GetMediaStream()` method to open a read stream of the chosen media item returned from the Take/Pick Photo/Video async methods.